app/requirements.txt
streamlit==1.21.0
pandas==2.1.0
sqlite3
Pillow==9.3.0
